Use uint instead of int in SmartHeader. Use uint instead of Height where Mempool height cannot happen.

+++ WalletWasabi.Tests/UnitTests/HashChainTests.cs
@@ -125,19 +125,19 @@ namespace WalletWasabi.Tests.UnitTests
 			// ASSERT PROPERTIES
 			Assert.Equal(0, hashChain.HashCount);
 			var hashesLeft = sameServerheight;
-			Assert.Equal(hashesLeft, hashChain.HashesLeft);
+			Assert.Equal((int)hashesLeft, hashChain.HashesLeft);
 			Assert.Equal(hashesLeft, hashChain.ServerTipHeight);
 			Assert.Null(hashChain.TipHash);
-			Assert.Equal(0, hashChain.TipHeight);
+			Assert.Equal(0u, hashChain.TipHeight);
 		}
 
 		private static void AssertEverythingDefault(HashChain hashChain)
 		{
 			Assert.Equal(0, hashChain.HashCount);
 			Assert.Equal(0, hashChain.HashesLeft);
-			Assert.Equal(0, hashChain.ServerTipHeight);
+			Assert.Equal(0u, hashChain.ServerTipHeight);
 			Assert.Null(hashChain.TipHash);
-			Assert.Equal(0, hashChain.TipHeight);
+			Assert.Equal(0u, hashChain.TipHeight);
 		}
 	}
 }
